+++ tests/Generator/Fixtures/AllOfRef/Output-5.6/MyClass.php
@@ -189,6 +189,17 @@ class MyClass
     }
 
     /**
+     * Converts this object back to an stdClass that can be JSON-serialized
+     *
+     * @return \stdClass Converted object
+     */
+    public function toStdClass()
+    {
+        $array = $this->toArray();
+        return json_decode(json_encode($array));
+    }
+
+    /**
      * Validates an input array
      *
      * @param array|object $input Input data