+++ tests/Generator/Fixtures/Descriptions/Output-7.4/Baz.php
@@ -118,6 +118,17 @@ class Baz
     }
 
     /**
+     * Converts this object back to an stdClass that can be JSON-serialized
+     *
+     * @return \stdClass Converted object
+     */
+    public function toStdClass(): \stdClass
+    {
+        $array = $this->toArray();
+        return json_decode(json_encode($array));
+    }
+
+    /**
      * Validates an input array
      *
      * @param array|object $input Input data