+++ tests/Generator/Fixtures/NestedTypedArrayProperty/Output-8.4/Fio.php
@@ -120,6 +120,17 @@ class Fio
     }
 
     /**
+     * Converts this object back to an stdClass that can be JSON-serialized
+     *
+     * @return \stdClass Converted object
+     */
+    public function toStdClass(): \stdClass
+    {
+        $array = $this->toArray();
+        return json_decode(json_encode($array));
+    }
+
+    /**
      * Validates an input array
      *
      * @param array|object $input Input data