+++ tests/Generator/Fixtures/NoEnums/Output-7.4/MyClass.php
@@ -110,6 +110,17 @@ class MyClass
     }
 
     /**
+     * Converts this object back to an stdClass that can be JSON-serialized
+     *
+     * @return \stdClass Converted object
+     */
+    public function toStdClass(): \stdClass
+    {
+        $array = $this->toArray();
+        return json_decode(json_encode($array));
+    }
+
+    /**
      * Validates an input array
      *
      * @param array|object $input Input data