+++ tests/Generator/Fixtures/ReferencedUnion/Output-5.6/MyObject.php
@@ -131,6 +131,17 @@ class MyObject
     }
 
     /**
+     * Converts this object back to an stdClass that can be JSON-serialized
+     *
+     * @return \stdClass Converted object
+     */
+    public function toStdClass()
+    {
+        $array = $this->toArray();
+        return json_decode(json_encode($array));
+    }
+
+    /**
      * Validates an input array
      *
      * @param array|object $input Input data